# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ResumeAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The type of text to enhance
 */
export type EnhanceTextInputType =
  (typeof EnhanceTextInputType)[keyof typeof EnhanceTextInputType];

export const EnhanceTextInputType = {
  summary: "summary",
  experience: "experience",
  project: "project",
  cover: "cover",
} as const;

export interface EnhanceTextInput {
  /** The type of text to enhance */
  type: EnhanceTextInputType;
  /** Existing text or role context to base the enhancement on */
  context: string;
  /**
   * The user's job title or target role
   * @nullable
   */
  jobTitle?: string | null;
  /**
   * Existing text to improve (if any)
   * @nullable
   */
  existingText?: string | null;
}

export interface EnhanceTextResult {
  /** The AI-generated or enhanced text */
  text: string;
}

export interface SuggestSkillsInput {
  /** Job title or role to suggest skills for */
  role: string;
  /**
   * Comma-separated existing skills to avoid duplicating
   * @nullable
   */
  existingSkills?: string | null;
}

export interface SuggestSkillsResult {
  /** List of suggested skills */
  skills: string[];
}

export interface ErrorResponse {
  error: string;
}