# lib/api-zod/src/generated/types/enhanceTextInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ResumeAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnhanceTextInputType } from "./enhanceTextInputType";

export interface EnhanceTextInput {
  /** The type of text to enhance */
  type: EnhanceTextInputType;
  /** Existing text or role context to base the enhancement on */
  context: string;
  /**
   * The user's job title or target role
   * @nullable
   */
  jobTitle?: string | null;
  /**
   * Existing text to improve (if any)
   * @nullable
   */
  existingText?: string | null;
}