# lib/api-zod/src/generated/types/enhanceTextResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ResumeAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnhanceTextResult {
  /** The AI-generated or enhanced text */
  text: string;
}