# lib/api-zod/src/generated/types/suggestSkillsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ResumeAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestSkillsInput {
  /** Job title or role to suggest skills for */
  role: string;
  /**
   * Comma-separated existing skills to avoid duplicating
   * @nullable
   */
  existingSkills?: string | null;
}