# lib/api-zod/src/generated/types/suggestSkillsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ResumeAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestSkillsResult {
  /** List of suggested skills */
  skills: string[];
}